% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{BasicGridCreation}
\alias{BasicGridCreation}
\title{Do not both to write help function}
\usage{
BasicGridCreation(countingDomain, grid.pixelsize, areaCountDomain)
}
\description{
Do not both to write help function
}
