% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{BasicResultExtraction}
\alias{BasicResultExtraction}
\title{Transform covariates to new grid}
\usage{
BasicResultExtraction(
  pp.res,
  inlaPrepList,
  projgrid,
  use.covariates = TRUE,
  covariate.fitting = "quadratic",
  additional.iid.term = TRUE,
  covariateValues,
  logicalGridPointsInsideCountingDomain,
  nxy,
  extraNonlinear.covGridList
)
}
\arguments{
\item{pp.res}{inla object (being the output of running the inla function)}

\item{inlaPrepList}{list of data provided to the inla function  (being the output from either PrepareINLAFuncContLikApprox or PrepareINLAFunc)}

\item{projgrid}{inla.mesh.projector object (being the output from running inla.mesh.projector)}

\item{use.covariates}{Logical, indicating whether covariates are used or not (see description!)}

\item{covariate.fitting}{String, indicating how to model covariates. "linear", quadratic (default) or "linAndLog", or FALSE for no covariates}

\item{additional.iid.term}{Logical, indicating whether to include an additional iid (Gaussian) term in the latent field specification. FALSE is default}

\item{covariateValues, }{Matrix giving the covariate values on the grid}

\item{logicalGridPointsInsideCountingDomain}{Logical vector, indicating which of the grid elements are within the counting domain}

\item{nxy}{Numberic vector of size 2, giving the dimension in x- and y-direction for the grid}

\item{extraNonlinear.covGridList}{List with additional projection object and such for nonlinear covariate effect, when applicable}
}
\value{
List with several results requiring minimal interaction with inla object
}
\description{
Transform a covariate grid of class 'im' to a new grid of covariate values, along with only those values being within
a specified counting domain. Note: This function should also be used if covariates are not present to simplify
later computations.
}
\keyword{inla}
