% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{BasicResultExtractionGAM}
\alias{BasicResultExtractionGAM}
\title{Do not both to write help function}
\usage{
BasicResultExtractionGAM(
  GAMfit,
  imputedList,
  use.covariates = TRUE,
  covariate.fitting = "quadratic",
  logicalGridPointsInsideCountingDomain,
  gridList
)
}
\description{
Do not both to write help function
}
