% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{ComparePhotoCountsAndPredGAM}
\alias{ComparePhotoCountsAndPredGAM}
\title{Comparison of photos counts and predictions for GAM}
\usage{
ComparePhotoCountsAndPredGAM(
  posteriorevalFullPhoto,
  posteriorDistPhoto,
  posteriorevalFullTransect,
  posteriorDistTransect,
  areaPerSubSampPhotoMatrix,
  photoCounts,
  photoOrder,
  leaveOutTransect,
  inputVar,
  savingFolder,
  plot = TRUE
)
}
\arguments{
\item{posteriorevalFullPhoto}{Numeric vector of evaluation points used for all posteriorDistributions}

\item{posteriorDistPhoto}{Matrix where each column is the posterior distribution probability for one photo}

\item{posteriorevalFullTransect}{Numeric vector of evaluation points used for posteriorDistTransect}

\item{posteriorDistTransect}{Numeric vector of posterior distribution probabilitites for the whole transect in question}

\item{photoCounts}{Numeric vector with the observed photo counts (not used in the modelling)}

\item{photoOrder}{Numeric vector specifying the plotting order of the photos (smallest to largest original x-coordinate)}

\item{leaveOutTransect}{The transect in question}

\item{inputVar}{List of all input variables used in the original run}

\item{savingFolder}{String with the path for where to save these comparison results}
}
\value{
Nothing, just save the output to file
}
\description{
Computes basic summary statistics based on distribution
}
