% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{ComputePostPredDistforPredPhotos}
\alias{ComputePostPredDistforPredPhotos}
\title{Computing the posterior predictive distribution of total area counts in the location of the PredPhotots}
\usage{
ComputePostPredDistforPredPhotos(
  samp,
  thisMeshPoint,
  covAtPredLoc,
  areaPredPhotos,
  poisson.maxEvals,
  parallelize.numCores,
  parallelize.noSplits,
  covariate.fitting,
  mesh
)
}
\arguments{
\item{samp}{List with all posterior samples, one sample in each sublist (as obtained from inla.posterior.sample)}

\item{thisMeshPoint}{Numeric vector giving the mesh point corresponding to each of the pred photos}

\item{covAtPredLoc}{Numeric vector containing the covariate value at the location of the prediction photos}

\item{poisson.maxEvals}{Numeric giving the maximum number of poisson evaluations}
}
\value{
List containing the resulting posterior predicitve distribution for all PredPoints
}
\description{
Currently only to be used with a spatial model with a linear covariate and no additional iid term
}
\keyword{inla}
