% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{coo2sp}
\alias{coo2sp}
\title{Transform regular coordinates to SpatialPolygons}
\usage{
coo2sp(coo)
}
\arguments{
\item{coo}{Numeric matrix of dim 2, specifying the x- and y-coordinates to transform}
}
\description{
Specify a function for transforming from regular coordinates to SpatialPolygons as used in the sp-package.
Copied from INLAs SPDE-tutorial
}
\keyword{Coordinates}
\keyword{tranformation}
