% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateCreation.R
\name{CovariateCreationFunction}
\alias{CovariateCreationFunction}
\title{Function for creating a dense grid with all covariate values}
\usage{
CovariateCreationFunction(
  satelliteFile = system.file("extdata", "original",
    "reflectance_0.0025deg_grid_modis_20120328_1310.tif", package = "SealCoxProcess"),
  sealFile = system.file("extdata", "original", "WestIce2012.csv", package =
    "SealCoxProcess"),
  xkmRange = c(-65.33908, 67.99593),
  ykmRange = c(-103.87783, 85.35911),
  covariates.type,
  chunksPerDim = 5,
  noPixEachDim = 1000,
  expandBy = 5
)
}
\arguments{
\item{xkmRange}{Vector of size 2, corresponding to the x-range of the mesh gotton with the standard input variables}

\item{ykmRange}{Vector of size 2, corresponding to the y-range of the mesh gotton with the standard input variables}

\item{covariates.type}{String indicating the type of covariate to create (either "band1" or "band2")}

\item{chunksPerDim}{Numeric. The number of chunkgs to split the data into when transforming the covariates}

\item{noPixEachDim}{Numeric. The number of pixels to use per dimension.}

\item{expandBy}{Numeric, indicating how many pixels the chunks should overlap with.}
}
\value{
Returns an image objects (which should be saved to file)
}
\description{
Transforms a tif file with lonlat values to a dense grid with the same coordinate system as the seal data (created with the lb2xykm function),
with mean values imputed where satellite image value is unknown. The output is a spatstat::im variable.
}
