% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{covAtNewGrid}
\alias{covAtNewGrid}
\title{Transform covariates to new grid}
\usage{
covAtNewGrid(
  use.covariates,
  covGrid,
  gridvalX,
  gridvalY,
  modelledGridVal,
  logicalGridPointsInsideCountingDomain
)
}
\arguments{
\item{use.covariates}{Logical, indicating whether covariates are used or not (see description!)}

\item{covGrid, }{im object representing the covariate values on a dense grid where the counts live.}

\item{gridvalX}{Numeric vector with the x-values for the grid}

\item{gridvalY}{Numeric vecotr with the y-values for the grid}

\item{modelledGridVal}{Matrix with same dimension as the new grid, where 1 indicates that the grid point is modelled, NA means it is outside the modelling domain.}

\item{logicalGridPointsInsideCountingDomain}{Logical vector, indicating which of the grid elements are within the counting domain}
}
\value{
List with the covariate values at the new grid (with respectively NA outside the modelling domain and outside the coutning domain)
}
\description{
Transform a covariate grid of class 'im' to a new grid of covariate values, along with only those values being within
a specified counting domain. Note: This function should also be used if covariates are not present to simplify
later computations.
}
\keyword{inla}
