% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{covGridCreationNonlinear}
\alias{covGridCreationNonlinear}
\title{Creating the grid for the nonlinear covariates}
\usage{
covGridCreationNonlinear(covMesh, covariateValues)
}
\arguments{
\item{covMesh}{Mesh object representing the mesh for the covariate when applicable, being the output from inla.mesh.1d}

\item{covariateValues, }{Matrix giving the covariate values on the grid}
}
\value{
List with the projection object and which values they represent
}
\description{
Extracts the nonlinear effect of a given set of covariate values
}
\keyword{inla}
