% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{CreateCountDomainPolygon}
\alias{CreateCountDomainPolygon}
\title{Create counting domain polygon}
\usage{
CreateCountDomainPolygon(
  transectStartCoordX,
  transectStartCoordY,
  transectEndCoordX,
  transectEndCoordY,
  coordPhotoX,
  coordPhotoY,
  photoWidth,
  photoHeight,
  transectYSpan = 1.5 * 1.852,
  transectXSpan = 0,
  theseTransInCountDomain = length(transectStartCoordX),
  cornerpointPolyInitializer = T
)
}
\arguments{
\item{transectStartCoordX}{Numeric vector with X-coordinate for which each transect is starting, in km}

\item{transectStartCoordY}{Numeric vector with Y-coordinate for which each transect is starting, in km}

\item{transectEndCoordX}{Numeric vector with X-coordinate for which each transect is ending, in km}

\item{transectEndCoordY}{Numeric vector with X-coordinate for which each transect is ending, in km}

\item{coordPhotoX}{Numeric vector with X-coordinate for the center point of each photo, in kilometers}

\item{coordPhotoY}{Numeric vector with Y-coordinate for the center point of each photo, in kilometers}

\item{photoWidth}{Numeric vector with the width of each photo, in km}

\item{photoHeight}{Numeric vector with the height of each photo, in km}

\item{transectYSpan}{Numeric, specifying how far transects are spanning in Y-direction, given in kilometers (for this data set this is 1.5Nm = 1.5*1.852 km)}

\item{transectXSpan}{Numeric, specifying how far transects are spanning in X-direction, given in kilometers}

\item{theseTransInCountDomain}{Vector of the transects to include in the counting domain (should be adjecent transects to fully sense), default is all (=length(transectStartCoordX))}
}
\value{
A polygon of the counting domain given as a data.frame with the x- and y-coordinates of the complete polygon
}
\description{
Creates a polygon corresponding to the area where seals might live and we are going to count
}
\keyword{polygon}
