% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{DefToGlob}
\alias{DefToGlob}
\title{Default function values to global variables}
\usage{
DefToGlob(fun, noDefaultVal = NULL)
}
\arguments{
\item{fun}{The function, whose defualt arguments are to be set as global arguments}

\item{noDefaultVal}{Which "value" to use for the arguments without a default value}
}
\value{
Creates variables with the default values as global variables
}
\description{
Function for setting default function arguments as global variables
}
