% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{FullPostPredDistGAM}
\alias{FullPostPredDistGAM}
\title{Do not both to write help function}
\usage{
FullPostPredDistGAM(
  samp,
  parallelize.noSplits = parallelize.numCores,
  parallelize.numCores,
  tempFolder,
  gridList,
  predPhotoGridPointList,
  est.theta,
  subSampPerSamp,
  fam,
  delete.samp = T
)
}
\description{
Do not both to write help function
}
