% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMFittingFunc.R
\name{GAMFittingFunc}
\alias{GAMFittingFunc}
\title{Do not both to write help function}
\usage{
GAMFittingFunc(
  resultsBaseFolder = "./Results/GAM",
  sealPhotoDataFile = system.file("extdata", "processed", "OriginalSealsKm.rds", package
    = "SealCoxProcess"),
  sealTransectDataFile = system.file("extdata", "processed",
    "OigardTablesTransformed.rds", package = "SealCoxProcess"),
  satelliteDataFolder = system.file("extdata", "processed", package = "SealCoxProcess"),
  sealType = "hooded",
  covariate.fitting = "linear",
  covariates.type = "band1",
  additional_comment = "",
  grid.pixelsize = 0.2,
  leaveOutTransect = FALSE,
  transectAsCountDomain = FALSE,
  noSamp = 1000,
  subSampPerSamp = 5,
  results.CI.level = 0.95,
  parallelize.numCores = 10,
  parallelize.noSplits = 2 * parallelize.numCores,
  delete.temp = TRUE,
  save.data = TRUE,
  fam = "negbin",
  CVFold = list(0, 0, 0)
)
}
\description{
Do not both to write help function
}
