% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{GAMImpute}
\alias{GAMImpute}
\title{Do not both to write help function}
\usage{
GAMImpute(
  gridvalX,
  gridvalY,
  use.covariates,
  covGrid,
  GAMfit,
  logicalGridPointsInsideCountingDomain
)
}
\description{
Do not both to write help function
}
