% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{GAMImputePred}
\alias{GAMImputePred}
\title{Do not both to write help function}
\usage{
GAMImputePred(allX, allY, use.covariates, covGrid, GAMfit)
}
\description{
Do not both to write help function
}
