% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{getstr}
\alias{getstr}
\title{String extraction function}
\usage{
getstr(mystring, initial.character = "_", final.character = "_")
}
\arguments{
\item{mystring}{Character vector to extract from.}

\item{initial.character}{Character determining the starting point of extractions}

\item{final.character}{Character determining the end point of extractions}
}
\value{
snippet
}
\description{
Function extracting string between two specific characters, minor customization of this one
http://www.r-bloggers.com/how-to-extract-a-string-between-2-characters-in-r-and-sas/
}
