% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{GridCreation}
\alias{GridCreation}
\title{Grid creation}
\usage{
GridCreation(
  mesh,
  countingDomain,
  areaCountDomain,
  grid.pixelsize = 0.2,
  gridSpan = "mesh"
)
}
\arguments{
\item{mesh}{Mesh object, being the output from inla.mesh.2d}

\item{countingDomain}{data.frame containing x- and y-coordinates of the counting domain polygon}

\item{areaCountDomain}{Numeric giving the area of the counting domain (in km)}

\item{grid.pixelsize}{Numeric, denoting the size (in km, in both x- and y-direction) of each pixel of the grid being used}

\item{gridSpan}{String equal to either "mesh" or "countingDomain" specifying what area the grid should span.}
}
\value{
List with the X and Y-values for the grid, the mesh-to-grid projection object (output form inla.mesh.projector), and other related variables
}
\description{
Creating a grid spanning the complete mesh with a specified pixel size and a projection from the mesh to the grid
}
\keyword{inla}
