% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{GridPointsInPredPhotos}
\alias{GridPointsInPredPhotos}
\title{Grid Points in prediction photos}
\usage{
GridPointsInPredPhotos(gridList, predPhotos)
}
\arguments{
\item{gridList}{List with grid objects, being the output from GridCreation}

\item{predPhotos}{data.frame containing x- and y-coordinates of the four corners of each of the pictures}
}
\value{
List with which gridpoints are within each of the pictures and how much their integrated intensity should be scaled to match the size of the picture
}
\description{
Finds which grid points are to be used to represent the value in each of the photos.
}
\keyword{inla}
