% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{GridPointsInPredPhotosGAM}
\alias{GridPointsInPredPhotosGAM}
\title{Do not both to write help function}
\usage{
GridPointsInPredPhotosGAM(gridList, predPhotos, grid.pixelsize)
}
\description{
Do not both to write help function
}
