% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{IndepMeshCreation}
\alias{IndepMeshCreation}
\title{Basic data independent mesh creation}
\usage{
IndepMeshCreation(
  rectangleCentersX,
  rectangleCentersY,
  countingDomainExpanded,
  convHullVar.convex = -0.15,
  convHullVar.concave = convHullVar.convex,
  convHullVar.resolution = c(120, 120),
  meshVar.max.edge = c(1, 10),
  meshVar.offset = c(6, 6),
  meshVar.cutoff = 0.3
)
}
\arguments{
\item{rectangleCentersX}{Vector with X-coordinate of the centerpoint of each of the obligatory rectangles}

\item{rectangleCentersY}{Vector with X-coordinate of the centerpoint of each of the obligatory rectangles}

\item{convHullVar.convex}{convex parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{convHullVar.concave}{concave parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{convHullVar.resolution}{resolution parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{meshVar.max.edge}{max.edge parameter of inla.mesh.2d. Smaller values gives smaller triangles outside triangles. See ?inla.mesh.2d for details}

\item{meshVar.offset}{offset parameter of inla.mesh.2d. See ?inla.mesh.2d for details}

\item{meshVar.cutoff}{cutoff parameter of inla.mesh.2d. See ?inla.mesh.2d for details}

\item{rectangleWidth}{Vector with the width of each of the obligatory rectangles}

\item{rectangleHeight}{Vector with the height of each of the obligatory rectanglges}
}
\value{
An INLA mesh object
}
\description{
Creates a 2D mesh with the inla.mesh.2d function with an inner and an outer domain with different triangle density.
The outer is defined as a larger nonconvex hull of the rectangleCenters while the inner is defined based on a slightly expanded counting domain
}
\keyword{mesh}
