% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonProcessFunction.R
\name{INLAPPSealsContLikApprox}
\alias{INLAPPSealsContLikApprox}
\title{Seal counting using INLA and continuous field Poisson process formulation (work in progress)}
\usage{
INLAPPSealsContLikApprox(
 
    resultsBaseFolder = "/nr/samba/user/jullum/Prosjekter/FRINATEK-Point/PointProcessesInINLA/Results/PoissonContLikApprox",
  sealPhotoDataFile = "/nr/project/stat/PointProcess/Data/Seals/OriginalSealsKm.rds",
 
    sealTransectDataFile = "/nr/project/stat/PointProcess/Data/Seals/OigardTablesTransformed.rds",
  satelliteDataFolder = "/nr/project/stat/PointProcess/Data/Seals/Satellite/",
  sealType = "hooded",
  spatial = TRUE,
  covariate.fitting = "quadratic",
  covariates.type = "band1",
  additional.iid.term = FALSE,
  observationMethod = "randomWithinPhoto",
  intPointMethod = "meshPoints",
  intPointWeightMethod = "Voronoi",
  noSamp = 5000,
  convHullVar.convex = -0.15,
  convHullVar.concave = convHullVar.convex,
  convHullVar.resolution = c(120, 120),
  meshVar.max.edge = c(2, 10),
  meshVar.offset = 6,
  meshVar.cutoff = 0.2,
  Matern.alpha = 2,
  grid.pixelsize = 0.2,
  INLA.theta.startval = NULL,
  INLA.verbose = FALSE,
  parallelize.numCores = 2,
  parallelize.noSplits = parallelize.numCores,
  poisson.maxEvals = 5 * 10^5,
  results.CI.level = 0.95,
  additional_comment = "",
  save.data = TRUE,
  delete.temp = TRUE,
  testing = FALSE
)
}
\arguments{
\item{resultsBaseFolder}{String, indicating the path to the base folder where all results are to be stored}

\item{sealPhotoDataFile}{String, indicating the file where seal photo data are stored}

\item{sealTransectDataFile}{String, indicating the file where seal transect data are stored}

\item{satelliteDataFolder}{String, indicating the path to the folder where the satellite data are stored (in files on the form "cov_grid_[covariates.type].rds")}

\item{sealType}{String, indicating which seal type to model "hooded" (default as it is quicker), or "harps"}

\item{spatial}{Logical indicating whether a spatial spde model should be used}

\item{covariate.fitting}{String, indicating how to model covariates. "linear", quadratic (default) or "linAndLog", or FALSE for no covariates}

\item{covariates.type}{String equal to "band1" or "band2" indicating which of the bands from the satellite data should be used.}

\item{additional.iid.term}{Logical, indicating whether to include an additional iid (Gaussian) term in the latent field specification. FALSE is default}

\item{observationMethod}{String, indicating which method to use when defining the locations. One of "asIs", "repeatCenter" or "randomWithinPhoto".
"asIs" takes the total number of counts in the center of each photo.
"repeatedCenter" repates the center location the same number of times as there are observations
"randomWithinPhoto" samples the location uniformly within each photo.}

\item{intPointMethod}{String, indicating which method to use when defining the integration points. One of "Fabian" or "mesh".
"Fabian" creates a new mesh for the purpose of selecting the integration points, located at the border of observation domain.
"meshPoints" uses the original mesh nodes as integration points.}

\item{intPointWeightMethod}{String, indicating which method to use when defining the weights for the integration points. One of "Fabian" or "Voronoi".
"Fabian" gives 1/3 of the area of each triangle as weight to the corner points (using the new mesh points giving triangles either completely insside or outside of the observation domain)
"Voronoi" uses the area of the Voronoi tiles (inside the observation domain) at each mesh node (with either integration point method) as the weight for each integration point.}

\item{noSamp}{Numeric, indicating the number of samples from posterior model to use when computing the posterior predictive distribution. 5000 (default) typically sufficient.}

\item{convHullVar.convex}{Numeric, corrresponding to the convex parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{convHullVar.concave}{Numeric, corrresponding to the concave parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{convHullVar.resolution}{Numeric vector of length 2, corrresponding to the resolution parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{meshVar.max.edge}{Numeric, corrresponding to the max.edge parameter of inla.mesh.2d. Smaller values gives smaller triangles outside triangles. See ?inla.mesh.2d for details}

\item{meshVar.offset}{Numeric, corrresponding to the offset parameter of inla.mesh.2d. See ?inla.mesh.2d for details}

\item{meshVar.cutoff}{Numeric, corrresponding to the cutoff parameter of inla.mesh.2d. See ?inla.mesh.2d for details}

\item{Matern.alpha}{Numeric, corresponding to the alpha parameter in the Matern covariance function (2 is the default)}

\item{grid.pixelsize}{Numeric, denoting the size (in km, in both x- and y-direction) of each pixel of the grid being used}

\item{INLA.theta.startval}{List containing the start values for the theta parameter in the INLA run. (NULL indicates that automatic start values should be used, and is the default)}

\item{INLA.verbose}{Logical, indicating whether verbose printing should be used when running the actual inla function}

\item{parallelize.numCores}{Numeric, corresponding to the number of cores any parallelization should be run at}

\item{parallelize.noSplits}{Numeric, deciding how many sublists samp should be splitted into. Should be a multiple of parallelize.numCores for the highest efficiency. The larger number the less memory is used (and longer time).}

\item{poisson.maxEvals}{Numeric, corresponding to maximum number of points the Poisson distribution should be evaluated at (a much smaller number is typically used)}

\item{results.CI.level}{Numeric, denoting the confidence/credibility degree to use in the final credibility interval for the total number of counts}

\item{additional_comment}{String, where any comments related to the actual run can be given}

\item{save.data}{Logical, indicating whether input variables, data and results should be saved (TRUE is default)}

\item{delete.temp}{Logical, indicating whether temporary stored samples and associated eta grids should be deleted}

\item{testing}{Logical, indicating whether the testing parts of this function should be used}
}
\value{
Nothing really, but saves results in the resultsBaseFolder
}
\description{
Approximates a Log-Gaussian Cox process by a continuous field approximation (Simpson et. al (2016)), and fits seal counts from transects to it.
Samples from the the fitted model to compute the posterior predictive distribution of the total number of seals within an
area which the transects are spanning
}
\keyword{Poisson-regression}
\keyword{cox-process}
\keyword{inla}
