% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{int.polygon}
\alias{int.polygon}
\title{Integration points for polygons inside an inla.mesh}
\usage{
int.polygon(mesh, loc, group = NULL)
}
\arguments{
\item{mesh}{An inla.mesh object}

\item{loc}{Locations defining the polygons}

\item{group}{If loc defines multiple polygons then this is the ID of the group for each location in loc}
}
\description{
Function for finding integration points for a given mesh and locations defining the polygons where observations are being made.
Created by Fabian Bachl.
}
\author{
Fabian E. Bachl <\email{f.e.bachl@bath.ac.uk}>
}
