% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{IntegrateVectorizedGrids}
\alias{IntegrateVectorizedGrids}
\title{Integrate over sampled filed on a grid}
\usage{
IntegrateVectorizedGrids(
  arrayGrid,
  logicalGridPointsInsideCountingDomain,
  truePixelSize,
  scaleArea
)
}
\arguments{
\item{arrayGrid}{Matrix where the first dimension gives the complete vectorized (1-dimensional) grid ( c() applied to the matrix of grid values) and teh second dimension corresponds to differnet samples}

\item{logicalGridPointsInsideCountingDomain}{Logical vector, indicating which of the grid elements are within the counting domain}

\item{truePixelSize}{Numeric vector of dim 2 specifying the size of each pixel (in x- and y-direction)}

\item{scaleArea}{Numeric specifying how much the resulting integreated field should be adjust to correct for the inaccurate size of the counting domain when gridding}
}
\value{
Numeric vector with the expected number of Poisson distributed counts given each of the sampled fields
}
\description{
Integrate over a sampled field given as a matrix within a specified counting region
}
\keyword{inla}
