% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{lapplyIntegrateVectorizedGrids}
\alias{lapplyIntegrateVectorizedGrids}
\title{Wrapper for IntegrateVectorizedGrids}
\usage{
lapplyIntegrateVectorizedGrids(listi, arrayGrid, truePixelSize)
}
\arguments{
\item{listi}{list with logicalGridPointsInsideCountingDomain and scaleArea stuff}

\item{arrayGrid}{Matrix where the first dimension gives the complete vectorized (1-dimensional) grid ( c() applied to the matrix of grid values) and teh second dimension corresponds to differnet samples}

\item{truePixelSize}{Numeric vector of dim 2 specifying the size of each pixel (in x- and y-direction)}
}
\value{
Numeric vector with the expected number of Poisson distributed counts given each of the sampled fields
}
\description{
Integrate over a sampled field given as a matrix within a specified counting region
}
\keyword{inla}
