% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{lapplymeanPostPredDistGAMBoth}
\alias{lapplymeanPostPredDistGAMBoth}
\title{Wrapper for samplePostPredDistGAM}
\usage{
lapplymeanPostPredDistGAMBoth(
  listi,
  arrayGrid,
  truePixelSize,
  allPhotosinGrid = rep(TRUE, nrow(arrayGrid))
)
}
\description{
Do not bother...
}
