% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{lapplysamplePostPredDistGAM}
\alias{lapplysamplePostPredDistGAM}
\title{Wrapper for samplePostPredDistGAM}
\usage{
lapplysamplePostPredDistGAM(
  listi,
  arrayGrid,
  truePixelSize,
  allPhotosinGrid = rep(TRUE, nrow(arrayGrid)),
  est.theta,
  subSampPerSamp
)
}
\description{
Do not bother...
}
