% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{lapplysamplePostPredDistGAMPoisson}
\alias{lapplysamplePostPredDistGAMPoisson}
\title{Wrapper for samplePostPredDistGAM}
\usage{
lapplysamplePostPredDistGAMPoisson(
  listi,
  arrayGrid,
  truePixelSize,
  allPhotosinGrid = rep(TRUE, nrow(arrayGrid)),
  subSampPerSamp
)
}
\description{
Do not bother...
}
