% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{lb2xykm}
\alias{lb2xykm}
\title{Contributed by Tor Arne Øigård}
\usage{
lb2xykm(lon, lat, lon0, lat0)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{lon0}{longitude base value}

\item{lat0}{latitude base value}
}
\description{
Contributed by Tor Arne Øigård
}
