% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{MeanNegbinDist}
\alias{MeanNegbinDist}
\title{Probability distribution computation for a mean of Negative binomial counts}
\usage{
MeanNegbinDist(eval, muvec, est.theta)
}
\arguments{
\item{eval}{Numeric vector with points where the distribution is evaluated}

\item{muvec}{Numeric vector with lambda values (characterizing the Negative binomial counts)}

\item{est.theta}{The fixed theta parameter}
}
\value{
Numeric vector with the expected number of Negative binomial distributed counts given each of the sampled fields
}
\description{
Computing the probability distribution for the mean of a set of Negative binomial distributed counts
}
\details{
Note: The distribution is not normalized
}
