% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{MeanPoissonDistMat}
\alias{MeanPoissonDistMat}
\title{Probability distribution computation for a mean of Poisson counts with a matrix of lambdas}
\usage{
MeanPoissonDistMat(eval, lambdaMat)
}
\arguments{
\item{eval}{Numeric vector with points where the distribution is evaluated}

\item{lambdavec}{Numeric vector with lambda values (characterizing the Poisson counts)}
}
\value{
Numeric vector with the expected number of Poisson distributed counts given each of the sampled fields
}
\description{
Computing the probability distribution for the mean of a set of Poisson distributed counts
}
\details{
Note: The distribution is not normalized
}
\keyword{inla}
