% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{MergePolygons}
\alias{MergePolygons}
\title{Merge polygons}
\usage{
MergePolygons(photos, plot = F)
}
\arguments{
\item{photos}{Data frame with 4 columns, the x-left, x-right, y-bottom, y-top coordinates of the photos}
}
\value{
A data frame with the merged polygons of the same format as the PolySet of the PBSmapping package
}
\description{
Merge several polygons into their union
}
\keyword{polygon}
