% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{MeshCreationMatchingRectangles}
\alias{MeshCreationMatchingRectangles}
\title{Mesh creation for rectangle matching}
\usage{
MeshCreationMatchingRectangles(
  rectangleCentersX,
  rectangleCentersY,
  rectangleWidth,
  rectangleHeight,
  convHullVar.convex = -0.15,
  convHullVar.concave = convHullVar.convex,
  convHullVar.resolution = c(120, 120),
  meshVar.max.edge = c(2, 10),
  meshVar.offset = 6,
  meshVar.cutoff = 0.2,
  y.cutoff.boundary = NULL,
  countingDomainExpanded
)
}
\arguments{
\item{rectangleCentersX}{Vector with X-coordinate of the centerpoint of each of the obligatory rectangles}

\item{rectangleCentersY}{Vector with X-coordinate of the centerpoint of each of the obligatory rectangles}

\item{rectangleWidth}{Vector with the width of each of the obligatory rectangles}

\item{rectangleHeight}{Vector with the height of each of the obligatory rectanglges}

\item{convHullVar.convex}{convex parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{convHullVar.concave}{concave parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{convHullVar.resolution}{resolution parameter of inla.nonvonvex.hull. See ?inla.nonconvex.hull for details}

\item{meshVar.max.edge}{max.edge parameter of inla.mesh.2d. Smaller values gives smaller triangles outside triangles. See ?inla.mesh.2d for details}

\item{meshVar.offset}{offset parameter of inla.mesh.2d. See ?inla.mesh.2d for details}

\item{meshVar.cutoff}{vector with cutoff parameters of inla.mesh.2d. See ?inla.mesh.2d for details}

\item{y.cutoff.boundary}{numeric vector deciding the y-values which should be the coundary for the  different cutoff-values. NULL means the first is applied everywhere}

\item{countingDomainExpanded}{Expanded coutning domain used to define the inner boundary for the mesh}
}
\value{
An INLA mesh object
}
\description{
Creates a 2D mesh with the inla.mesh.2d function which have a Voronoi triangulation which matches rectanglges given as input.
}
\keyword{mesh}
