% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PhotoPostPredDistGAM}
\alias{PhotoPostPredDistGAM}
\title{Do not both to write help function}
\usage{
PhotoPostPredDistGAM(
  samp,
  parallelize.noSplits = parallelize.numCores,
  parallelize.numCores,
  tempFolder,
  gridList,
  predPhotoGridPointList,
  est.theta,
  allPhotosinGrid,
  subSampPerSamp,
  fam
)
}
\description{
Do not both to write help function
}
