% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PhotoPostPredDistold}
\alias{PhotoPostPredDistold}
\title{Computing the posterior predictive distribution for single photos and the sums of these photos in addition to the regular counting domain}
\usage{
PhotoPostPredDistold(
  samp,
  spatial,
  parallelize.noSplits = parallelize.numCores,
  parallelize.numCores,
  tempFolder,
  use.covariates,
  additional.iid.term,
  covariate.fitting,
  gridList,
  covGridList,
  nxy,
  poisson.maxEvals,
  noMeshPoints,
  extraNonlinear.covGridList,
  predPhotoGridPointList
)
}
\arguments{
\item{samp}{List with all posterior samples, one sample in each sublist (as obtained from inla.posterior.sample)}

\item{spatial}{Logical, indicating whether the model fitted includes a spatial spde term}

\item{parallelize.noSplits}{Numeric, deciding how many sublists samp should be splitted into. Should be a multiple of parallelize.numCores for the highest efficiency. The larger number the less memory is used (and longer time)}

\item{parallelize.numCores}{Numeric, corresponding to the number of cores any parallelization should be run at}

\item{tempFolder}{Path to the folder where temporary files (samp subfiles and gridded eta subfiles)}

\item{use.covariates}{Logical, indicating whether covariates are used or not (see description!)}

\item{additional.iid.term}{Logical, indicating whether to include an additional iid (Gaussian) term in the latent field specification. FALSE is default}

\item{covariate.fitting}{String, indicating how to model covariates. "linear", quadratic (default) or "linAndLog", or FALSE for no covariates}

\item{gridList}{List containing information about the grid, being the output of the function GridCreation}

\item{covGridList}{List containing information about the covariates at the grid (also when use.covariates = FALSE), being the output of the function covAtNewGrid}

\item{nxy}{Numberic vector of size 2, giving the dimension in x- and y-direction for the grid}

\item{poisson.maxEvals}{Numeric, corresponding to maximum number of points the Poisson distribution should be evaluated at (a much smaller number is typically used)}

\item{noMeshPoints}{Numeric, corresponding to the number of points in the mesh being used}

\item{extraNonlinear.covGridList}{List of additional projection objects related to the nonlinear covariate effect when applicable}

\item{predPhotoGridPointList}{List of which grid point which define the photos when sampling}
}
\value{
List containing the resulting posterior predicitve distribution for all evaluation points (also returned) in addition to the mean field and sd field computed from the samples)
}
\description{
As for the ComputePostPredDist, but with the additional stuff for doing it also per photo
}
\keyword{inla}
