% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{postPredDistGAMFunc}
\alias{postPredDistGAMFunc}
\title{Do not both to write help function}
\usage{
postPredDistGAMFunc(GAMfit, noSamp, imputedList, allPhotosinGrid)
}
\description{
Do not both to write help function
}
