% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PrepareINLAFunc}
\alias{PrepareINLAFunc}
\title{Prepare for running INLA functio}
\usage{
PrepareINLAFunc(
  mesh,
  covMesh,
  obsLoc,
  y.pp,
  e.pp,
  covGrid,
  spatial = TRUE,
  covariate.fitting = "quadratic",
  additional.iid.term = FALSE,
  Matern.alpha = 2,
  covariates.type = "band1",
  INLA.theta.startval = NULL,
  INLA.constr = TRUE
)
}
\arguments{
\item{mesh}{Mesh object, being the output from inla.mesh.2d}

\item{covMesh}{Mesh object representing the mesh for the covariate when applicable, being the output from inla.mesh.1d}

\item{obsLoc}{Matrix with at least two columns specifying, respectively, the x- and y-coordinates of the observations}

\item{y.pp}{Numeric vector, indicating number of counts at each of the locations}

\item{e.pp}{Numeric vector, indicating the weight to be assigned to each location}

\item{covGrid, }{im object representing the covariate values on a dense grid where the counts live.}

\item{spatial}{Logical indicating whether a spatial spde model should be used}

\item{covariate.fitting}{String, indicating how to model covariates. "no", "linear", "quadratic" (default), "linAndLog" or "nonlinear"}

\item{additional.iid.term}{Logical, indicating whether to include an additional iid (Gaussian) term in the latent field specification. FALSE is default}

\item{Matern.alpha}{Numeric, corresponding to the alpha parameter in the Matern covariance function (2 is the default)}

\item{covariates.type}{String equal to "band1" or "band2" indicating which of the bands from the satellite data should be used.}

\item{INLA.theta.startval}{List containing the start values for the theta parameter in the INLA run. (NULL indicates that automatic start values should be used, and is the default)}
}
\value{
List with all variables necessary to run the INLA function, in addition to the spde object
}
\description{
Gathers and stacks the input to the INLA function
}
\keyword{inla}
