% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PrepareINLAFuncContLikApprox}
\alias{PrepareINLAFuncContLikApprox}
\title{Prepare for running INLA function with continuous likelihood approximation}
\usage{
PrepareINLAFuncContLikApprox(
  mesh,
  intPoints,
  obsPoints,
  covGrid,
  spatial = TRUE,
  use.covariates = TRUE,
  covariate.fitting = "quadratic",
  additional.iid.term = FALSE,
  Matern.alpha = 2,
  covariates.type = "band1",
  INLA.theta.startval = NULL
)
}
\arguments{
\item{mesh}{Mesh object, being the output from inla.mesh.2d}

\item{intPoints}{Data frame with the integration point locations and their associated weights}

\item{obsPoints}{Data frame with the point pattern locations (locations with at least one seal observed)}

\item{covGrid, }{im object representing the covariate values on a dense grid where the counts live.}

\item{spatial}{Logical indicating whether a spatial spde model should be used}

\item{use.covariates}{Logical, indicating whether covariates should be used in the fitting (default is true)}

\item{covariate.fitting}{String, indicating how to model covariates. "linear", quadratic (default) or "linAndLog", or FALSE for no covariates}

\item{additional.iid.term}{Logical, indicating whether to include an additional iid (Gaussian) term in the latent field specification. FALSE is default}

\item{Matern.alpha}{Numeric, corresponding to the alpha parameter in the Matern covariance function (2 is the default)}

\item{covariates.type}{String equal to "band1" or "band2" indicating which of the bands from the satellite data should be used.}

\item{INLA.theta.startval}{List containing the start values for the theta parameter in the INLA run. (NULL indicates that automatic start values should be used, and is the default)}
}
\value{
List with all variables necessary to run the INLA function, in addition to the spde object
}
\description{
Gathers and stacks the input to the INLA function
}
\keyword{inla}
