% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{rnbinomSum}
\alias{rnbinomSum}
\title{Reversed negative binomial sampler}
\usage{
rnbinomSum(muVec, est.theta, subSampPerSamp)
}
\arguments{
\item{muVec}{vector or means}

\item{est.theta}{numeric theta value}
}
\value{
one sample for every value of mu
}
\description{
Reversed negative binomial sampler
}
