% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{rpoisPredPhoto}
\alias{rpoisPredPhoto}
\title{Reversed poisson sampler for pred photo}
\usage{
rpoisPredPhoto(muVec, subSampPerSamp)
}
\arguments{
\item{muVec}{vector or means}

\item{subSampPerSamp}{...}
}
\value{
one sample for every value of mu
}
\description{
Reversed poisson sampler for pred photo
}
