% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{rpoisSum}
\alias{rpoisSum}
\title{Reversed poisson sampler}
\usage{
rpoisSum(muVec, subSampPerSamp)
}
\arguments{
\item{muVec}{vector or means}

\item{subSampPerSamp}{...}
}
\value{
one sample for every value of mu
}
\description{
Reversed poisson sampler
}
