% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{samplePostPredDistGAM}
\alias{samplePostPredDistGAM}
\title{Sample postpredDistGAM}
\usage{
samplePostPredDistGAM(
  arrayGrid,
  logicalGridPointsInsideCountingDomain,
  truePixelSize,
  scaleArea,
  est.theta,
  subSampPerSamp
)
}
\arguments{
\item{arrayGrid}{Matrix where the first dimension gives the complete vectorized (1-dimensional) grid ( c() applied to the matrix of grid values) and teh second dimension corresponds to differnet samples}

\item{logicalGridPointsInsideCountingDomain}{Logical vector, indicating which of the grid elements are within the counting domain}

\item{truePixelSize}{Numeric vector of dim 2 specifying the size of each pixel (in x- and y-direction)}

\item{scaleArea}{Numeric specifying how much the resulting integreated field should be adjust to correct for the inaccurate size of the counting domain when gridding}

\item{est.theta}{numeric theta value}
}
\value{
Samples from predictive disttribution of fitted GAM model
}
\description{
Sample from the posterior predictive distribution for the complete domain
}
\keyword{inla}
