% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{samplePostPredDistGAMPoissonPredPhoto}
\alias{samplePostPredDistGAMPoissonPredPhoto}
\title{Sample samplePostPredDistGAMPoisson}
\usage{
samplePostPredDistGAMPoissonPredPhoto(
  arrayGrid,
  subSampPerSamp,
  areaPredPhotos
)
}
\description{
DO not bother
}
