% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{samplePostPredDistGAMPredPhoto}
\alias{samplePostPredDistGAMPredPhoto}
\title{Sample postpredDistGAM}
\usage{
samplePostPredDistGAMPredPhoto(
  arrayGrid,
  est.theta,
  subSampPerSamp,
  areaPredPhotos
)
}
\description{
DO not bother
}
