% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{SummaryPlotFuncGAM}
\alias{SummaryPlotFuncGAM}
\title{Do not both to write help function}
\usage{
SummaryPlotFuncGAM(
  covariatesplot = TRUE,
  summaryplot = TRUE,
  savingFolder,
  sealPhotoDataFile,
  sealTransectDataFile,
  dataList,
  orgPhotos,
  modPhotos,
  results.CI.level = 0.95,
  gridList,
  finalResList,
  countingDomain,
  logicalGridPointsInsideCountingDomain,
  covNewGridval,
  GAMfit,
  sealType,
  use.covariates,
  covariates.type,
  covariate.fitting,
  grid.pixelsize,
  parallelize.noSplits,
  parallelize.numCores,
  noSamp,
  subSampPerSamp,
  time,
  testing,
  comment,
  leaveOutTransect,
  fam
)
}
\description{
Do not both to write help function
}
