% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{SummaryStat}
\alias{SummaryStat}
\title{Posterior summary statistics}
\usage{
SummaryStat(evalPoints, dist, results.CI.level = 0.95, posterior = TRUE)
}
\arguments{
\item{evalPoints}{Numeric vector of evaluation points for a distribution}

\item{dist}{Numeric vector with the probability at the corresponding evaluation points}

\item{results.CI.level}{Numeric, denoting the confidence/credibility degree to use in the final credibility interval for the total number of counts}

\item{posterior}{Logical, indicating whether this is a posterior distribution for which the name posterior should be used on the output}
}
\value{
List with basic self-explanatory summary statistics
}
\description{
Computes basic summary statistics based on distribution
}
